[Feature] Account Net Worth Graph (#5037)
* show net worth graph for each account page

* add release notes

* hide filter button

* fix lint

* import ReactNode type

* find selected accounts based on accountId param

* remove breaks

* can toggle account page net worth graph

* Improves account balance history graph
Refactors the balance history graph to improve its data fetching and rendering.

Removes the `selectedAccounts` state and related logic from the Account component, simplifying its state management. The `accountId` is now passed directly to the BalanceHistoryGraph.

Moves the BalanceHistoryGraph component to the accounts directory and removes the redundant BalanceHistoryGraph in the sidebar.

Updates the queries to directly use the accountId for filtering transactions, leading to more efficient data retrieval.

* mv

* min width and height only for sidebar

* move toggle chart button to account menu

* add fill gradiant to chart

* fix maxWidth

* fix chart hover offset

* responsive graph aspect ratio

* auto hide chart on short view

* revert balance position

* auto hide chart on 15% height

* remove chart boolean from internal state

* Implement account-specific net worth chart preferences

Replaced global net worth chart visibility preference with account-specific preferences. Updated components to use the new dynamic preference keys and added a toggle option in the account menus for better user customization.

* lint

* fix account groups

* Update VRT

* bump

---------
