+++ packages/desktop-client/src/components/accounts/Header.tsx
@@ -1,6 +1,7 @@
 import React, {
   type ComponentProps,
   type ReactNode,
+  useEffect,
   useRef,
   useState,
 } from 'react';
@@ -42,6 +43,7 @@ import {
 
 import { type TableRef } from './Account';
 import { Balances } from './Balance';
+import { BalanceHistoryGraph } from './BalanceHistoryGraph';
 import { ReconcileMenu, ReconcilingMessage } from './Reconcile';
 
 import { AnimatedRefresh } from '@desktop-client/components/AnimatedRefresh';
@@ -55,6 +57,7 @@ import { useDateFormat } from '@desktop-client/hooks/useDateFormat';
 import { useLocale } from '@desktop-client/hooks/useLocale';
 import { useLocalPref } from '@desktop-client/hooks/useLocalPref';
 import { useSplitsExpanded } from '@desktop-client/hooks/useSplitsExpanded';
+import { useSyncedPref } from '@desktop-client/hooks/useSyncedPref';
 import { useSyncServerStatus } from '@desktop-client/hooks/useSyncServerStatus';
 
 type AccountHeaderProps = {
@@ -62,6 +65,7 @@ type AccountHeaderProps = {
   isNameEditable: boolean;
   workingHard: boolean;
   accountName: string;
+  accountId?: string;
   account?: AccountEntity;
   filterId?: SavedFilter;
   savedFilters: TransactionFilterEntity[];
@@ -137,6 +141,7 @@ export function AccountHeader({
   isNameEditable,
   workingHard,
   accountName,
+  accountId,
   account,
   filterId,
   savedFilters,
@@ -199,6 +204,10 @@ export function AccountHeader({
   const isUsingServer = syncServerStatus !== 'no-server';
   const isServerOffline = syncServerStatus === 'offline';
   const [_, setExpandSplitsPref] = useLocalPref('expand-splits');
+  const [showNetWorthChartPref, setShowNetWorthChartPref] = useSyncedPref(
+    `show-account-${accountId}-net-worth-chart`,
+  );
+  const showNetWorthChart = showNetWorthChartPref === 'true';
 
   const dateFormat = useDateFormat() || 'MM/dd/yyyy';
   const locale = useLocale();
@@ -222,6 +231,32 @@ export function AccountHeader({
       setExpandSplitsPref(!(splitsExpanded.state.mode === 'expand'));
     }
   }
+
+  const graphRef = useRef<HTMLDivElement>(null);
+  useEffect(() => {
+    const handleResize = () => {
+      const ele = graphRef.current;
+      if (!ele) return;
+      const clone = ele.cloneNode(true) as HTMLDivElement;
+      Object.assign(clone.style, {
+        visibility: 'hidden',
+        display: 'flex',
+      });
+      ele.after(clone);
+      if (clone.clientHeight < window.innerHeight * 0.15) {
+        setShowNetWorthChartPref('true');
+      } else {
+        setShowNetWorthChartPref('false');
+      }
+      clone.remove();
+    };
+
+    window.addEventListener('resize', handleResize);
+
+    return () => {
+      window.removeEventListener('resize', handleResize);
+    };
+  }, [setShowNetWorthChartPref]);
 
   useHotkeys(
     'ctrl+f, cmd+f, meta+f',
@@ -269,41 +304,63 @@ export function AccountHeader({
     <>
       <View style={{ ...styles.pageContent, paddingBottom: 10, flexShrink: 0 }}>
         <View
-          style={{ marginTop: 2, marginBottom: 10, alignItems: 'flex-start' }}
+          style={{
+            flexDirection: 'column',
+            marginTop: 2,
+            justifyContent: 'space-between',
+            gap: 10,
+          }}
         >
           <View
             style={{
-              flexDirection: 'row',
-              alignItems: 'center',
-              gap: 3,
+              flexGrow: 1,
+              alignItems: 'flex-start',
+              gap: 10,
             }}
           >
-            {!!account?.bank && (
-              <AccountSyncSidebar
+            <View
+              style={{
+                flexDirection: 'row',
+                alignItems: 'center',
+                gap: 3,
+              }}
+            >
+              {!!account?.bank && (
+                <AccountSyncSidebar
+                  account={account}
+                  failedAccounts={failedAccounts}
+                  accountsSyncing={accountsSyncing}
+                />
+              )}
+              <AccountNameField
                 account={account}
-                failedAccounts={failedAccounts}
-                accountsSyncing={accountsSyncing}
+                accountName={accountName}
+                isNameEditable={isNameEditable}
+                saveNameError={saveNameError}
+                onSaveName={onSaveName}
               />
-            )}
-            <AccountNameField
+            </View>
+
+            <Balances
+              balanceQuery={balanceQuery}
+              showExtraBalances={showExtraBalances}
+              onToggleExtraBalances={onToggleExtraBalances}
               account={account}
-              accountName={accountName}
-              isNameEditable={isNameEditable}
-              saveNameError={saveNameError}
-              onSaveName={onSaveName}
+              isFiltered={isFiltered}
+              filteredAmount={filteredAmount}
             />
           </View>
+
+          <BalanceHistoryGraph
+            ref={graphRef}
+            accountId={accountId}
+            style={{
+              height: 'calc(5vh + 5vw)',
+              margin: 0,
+              display: showNetWorthChart ? 'flex' : 'none',
+            }}
+          />
         </View>
-
-        <Balances
-          balanceQuery={balanceQuery}
-          showExtraBalances={showExtraBalances}
-          onToggleExtraBalances={onToggleExtraBalances}
-          account={account}
-          isFiltered={isFiltered}
-          filteredAmount={filteredAmount}
-        />
-
         <Stack
           spacing={2}
           direction="row"
@@ -484,6 +541,7 @@ export function AccountHeader({
                     <AccountMenu
                       account={account}
                       canSync={canSync}
+                      showNetWorthChart={showNetWorthChart}
                       canShowBalances={
                         canCalculateBalance ? canCalculateBalance() : false
                       }
@@ -523,6 +581,12 @@ export function AccountHeader({
                             ]
                           : []),
                         { name: 'export', text: t('Export') },
+                        {
+                          name: 'toggle-net-worth-chart',
+                          text: showNetWorthChart
+                            ? t('Hide balance chart')
+                            : t('Show balance chart'),
+                        },
                       ]}
                     />
                   </Dialog>
@@ -531,7 +595,6 @@ export function AccountHeader({
             </View>
           )}
         </Stack>
-
         {filterConditions?.length > 0 && (
           <FiltersStack
             conditions={filterConditions}
@@ -615,39 +678,38 @@ function AccountNameField({
     setEditingName(false);
   };
 
-  if (editingName) {
-    return (
-      <>
-        <InitialFocus>
-          <Input
-            defaultValue={accountName}
-            onEnter={handleSave}
-            onUpdate={handleSave}
-            onEscape={() => setEditingName(false)}
-            style={{
-              fontSize: 25,
-              fontWeight: 500,
-              marginTop: -3,
-              marginBottom: -4,
-              marginLeft: -6,
-              paddingTop: 2,
-              paddingBottom: 2,
-              width: Math.max(20, accountName.length) + 'ch',
-            }}
-          />
-        </InitialFocus>
-        {saveNameError && (
-          <View style={{ color: theme.warningText }}>{saveNameError}</View>
-        )}
-      </>
-    );
-  } else {
-    if (isNameEditable) {
-      return (
+  return (
+    <View style={{ flexShrink: 0, alignItems: 'center' }}>
+      {editingName ? (
+        <>
+          <InitialFocus>
+            <Input
+              defaultValue={accountName}
+              onEnter={handleSave}
+              onUpdate={handleSave}
+              onEscape={() => setEditingName(false)}
+              style={{
+                fontSize: 25,
+                fontWeight: 500,
+                marginTop: -3,
+                marginBottom: -4,
+                marginLeft: -6,
+                paddingTop: 2,
+                paddingBottom: 2,
+                width: Math.max(20, accountName.length) + 'ch',
+              }}
+            />
+          </InitialFocus>
+          {saveNameError && (
+            <View style={{ color: theme.warningText }}>{saveNameError}</View>
+          )}
+        </>
+      ) : (
         <View
           style={{
             flexDirection: 'row',
             alignItems: 'center',
+            whiteSpace: 'nowrap',
             gap: 3,
             '& .hover-visible': {
               opacity: 0,
@@ -672,46 +734,40 @@ function AccountNameField({
               : accountName}
           </View>
 
-          {account && (
-            <NotesButton
-              id={`account-${account.id}`}
-              defaultColor={theme.pageTextSubdued}
-            />
-          )}
-          <Button
-            variant="bare"
-            aria-label={t('Edit account name')}
-            className="hover-visible"
-            onPress={() => setEditingName(true)}
-          >
-            <SvgPencil1
-              style={{
-                width: 11,
-                height: 11,
-                color: theme.pageTextSubdued,
-              }}
-            />
-          </Button>
+          <View style={{ flexDirection: 'row', width: 50 }}>
+            {isNameEditable && account && (
+              <NotesButton
+                id={`account-${account.id}`}
+                defaultColor={theme.pageTextSubdued}
+              />
+            )}
+            {isNameEditable && (
+              <Button
+                variant="bare"
+                aria-label={t('Edit account name')}
+                className="hover-visible"
+                onPress={() => setEditingName(true)}
+              >
+                <SvgPencil1
+                  style={{
+                    width: 11,
+                    height: 11,
+                    color: theme.pageTextSubdued,
+                  }}
+                />
+              </Button>
+            )}
+          </View>
         </View>
-      );
-    } else {
-      return (
-        <View
-          style={{ fontSize: 25, fontWeight: 500, marginBottom: -1 }}
-          data-testid="account-name"
-        >
-          {account && account.closed
-            ? t('Closed: {{ accountName }}', { accountName })
-            : accountName}
-        </View>
-      );
-    }
-  }
+      )}
+    </View>
+  );
 }
 
 type AccountMenuProps = {
   account: AccountEntity;
   canSync: boolean;
+  showNetWorthChart: boolean;
   showBalances: boolean;
   canShowBalances: boolean;
   showCleared: boolean;
@@ -727,13 +783,15 @@ type AccountMenuProps = {
       | 'toggle-balance'
       | 'remove-sorting'
       | 'toggle-cleared'
-      | 'toggle-reconciled',
+      | 'toggle-reconciled'
+      | 'toggle-net-worth-chart',
   ) => void;
 };
 
 function AccountMenu({
   account,
   canSync,
+  showNetWorthChart,
   showBalances,
   canShowBalances,
   showCleared,
@@ -769,6 +827,12 @@ function AccountMenu({
               } as const,
             ]
           : []),
+        {
+          name: 'toggle-net-worth-chart',
+          text: showNetWorthChart
+            ? t('Hide balance chart')
+            : t('Show balance chart'),
+        },
         {
           name: 'toggle-cleared',
           text: showCleared
