+++ upcoming-release-notes/5037.md
@@ -1,0 +1,6 @@
+---
+category: Features
+authors: [passabilities]
+---
+
+Adds net worth graph for each account page.
